[CON-3918] BUG: Removed AddHttpClient registration of IRoatpApiClient as this is being duplicated during the structuremap container registration UseDefaultConventions().
Not a simple change to move around the ordering of the code and would potentially break the other calls to AddHttpClient as they would need to be seperately registered to interfaces also.

Now the _.WithDefaultConventions handles For<IRoatpApiClient, RoatpApiClient> and the call to AddHttpClient is responsible for injectng the appropiate HttpClient with correct config to the concrete RoatpApiClient.