Update to show .NET 6.0 and Visual Studio 2022
As @PeterDavidWilliams pointed out to me, the definition ID in the URL is linking to the wrong pipeline instead of the one for das-assessor-service, which might be why some builds weren't showing as failing when they should be. While looking at that, I noticed a few other places in the readme that could use some updates.